requirements.txt
python-dateutil==2.8.2
numpy==1.25.1
pandas==2.0.3
matplotlib==3.8.0
scikit-learn==1.3.0